Add files via upload
Explore the changing world of cybersecurity threats and weaknesses with our easy-to-use Security Vulnerabilities dataset. As technology becomes a bigger part of our daily lives, it’s more important than ever to find and understand security risks. This dataset is a helpful tool for anyone interested in digital security—whether you’re a cybersecurity professional, a data enthusiast, or a researcher. Use it to learn more about how to protect information and stay safe online.